refactor: move event mappings logic to their own dir
