+++ src/index.js
@@ -14,8 +14,21 @@
  */
 'use strict'
 const http = require('http')
-const url = require('url')
-const isType = require('type-is')
+const {
+  forwardRequestToNodeServer,
+  startServer,
+  makeResolver,
+  getSocketPath
+} = require('./transport')
+const {
+  getRandomString
+} = require('./utils')
+const {
+  getEventFnsBasedOnEventSource
+} = require('./event-mappings')
+const {
+  getEventSourceBasedOnEvent
+} = require('./event-mappings/utils')
 
 const currentLambdaInvoke = {}
 
@@ -26,211 +39,6 @@ function getCurrentLambdaInvoke () {
 function setCurrentLambdaInvoke ({ event, context }) {
   currentLambdaInvoke.event = event
   currentLambdaInvoke.context = context
-}
-
-function getPathWithQueryStringParams ({ event }) {
-  return url.format({
-    pathname: event.path,
-    query: event.multiValueQueryStringParameters
-  })
-}
-
-function getEventBody ({ event }) {
-  return Buffer.from(event.body, event.isBase64Encoded ? 'base64' : 'utf8')
-}
-
-function getContentType ({ contentTypeHeader }) {
-  // only compare mime type; ignore encoding part
-  return contentTypeHeader ? contentTypeHeader.split(';')[0] : ''
-}
-
-function isContentTypeBinaryMimeType ({ contentType, binaryMimeTypes }) {
-  return binaryMimeTypes.length > 0 && !!isType.is(contentType, binaryMimeTypes)
-}
-
-function mapEventToHttpRequest ({
-  event,
-  socketPath,
-  headers = {
-    ...event.multiValueHeaders
-  }
-}) {
-  return {
-    method: event.httpMethod,
-    path: getPathWithQueryStringParams({ event }),
-    headers,
-    socketPath
-    // protocol: `${headers['X-Forwarded-Proto']}:`,
-    // host: headers.Host,
-    // hostname: headers.Host, // Alias for host
-    // port: headers['X-Forwarded-Port']
-  }
-}
-
-function mapApiGatewayEventToHttpRequest ({ event, socketPath }) {
-  const httpRequest = mapEventToHttpRequest({ event, socketPath })
-
-  // NOTE: API Gateway is not setting Content-Length header on requests even when they have a body
-  if (event.body && !httpRequest.headers['Content-Length']) {
-    const body = getEventBody({ event })
-    httpRequest.headers['Content-Length'] = Buffer.byteLength(body)
-  }
-
-  return httpRequest
-}
-
-function mapAlbEventToHttpRequest ({ event, socketPath }) {
-  const httpRequest = mapEventToHttpRequest({ event, socketPath })
-
-  return httpRequest
-}
-
-function mapLambdaEdgeEventToHttpRequest ({ event, context, socketPath }) {
-  const httpRequest = mapEventToHttpRequest({ event, socketPath })
-
-  return httpRequest
-}
-
-function forwardResponse ({ server, response, resolver, responseFn }) {
-  return responseFn({ server, response, resolver })
-}
-
-function forwardResponseToApiGateway ({ server, response, resolver }) {
-  let buf = []
-
-  response
-    .on('data', (chunk) => buf.push(chunk))
-    .on('end', () => {
-      const bodyBuffer = Buffer.concat(buf)
-      const statusCode = response.statusCode
-      const headers = response.headers
-
-      // chunked transfer not currently supported by API Gateway
-      /* istanbul ignore else */
-      if (headers['transfer-encoding'] === 'chunked') {
-        delete headers['transfer-encoding']
-      }
-
-      const contentType = getContentType({
-        contentTypeHeader: headers['content-type']
-      })
-      const isBase64Encoded = isContentTypeBinaryMimeType({
-        contentType,
-        binaryMimeTypes: server._binaryMimeTypes
-      })
-      const body = bodyBuffer.toString(isBase64Encoded ? 'base64' : 'utf8')
-      const successResponse = {
-        statusCode,
-        body,
-        multiValueHeaders: headers,
-        isBase64Encoded
-      }
-
-      resolver.succeed({
-        response: successResponse
-      })
-    })
-}
-
-function forwardConnectionErrorResponseToApiGateway ({ error, resolver }) {
-  console.log('ERROR: aws-serverless-express connection error')
-  console.error(error)
-  const errorResponse = {
-    statusCode: 502, // "DNS resolution, TCP level errors, or actual HTTP parse errors" - https://nodejs.org/api/http.html#http_http_request_options_callback
-    body: '',
-    multiValueHeaders: {}
-  }
-
-  resolver.succeed({
-    response: errorResponse
-  })
-}
-
-function forwardLibraryErrorResponseToApiGateway ({ error, resolver }) {
-  console.log('ERROR: aws-serverless-express error')
-  console.error(error)
-  const errorResponse = {
-    statusCode: 500,
-    body: '',
-    multiValueHeaders: {}
-  }
-
-  resolver.succeed({
-    response: errorResponse
-  })
-}
-
-function getEventFnsBasedOnEventSource ({ eventSource }) {
-  switch (eventSource) {
-    case 'API_GATEWAY':
-      return {
-        request: mapApiGatewayEventToHttpRequest,
-        response: forwardResponseToApiGateway
-      }
-    case 'ALB':
-      return {
-        request: mapAlbEventToHttpRequest,
-        response: forwardResponseToApiGateway
-      }
-    case 'LAMBDA_EDGE':
-      return {
-        request: mapLambdaEdgeEventToHttpRequest,
-        response: forwardResponseToApiGateway
-      }
-    default:
-      return {
-        request: mapEventToHttpRequest,
-        response: forwardResponseToApiGateway
-      }
-  }
-}
-
-function forwardRequestToNodeServer ({
-  server,
-  event,
-  context,
-  resolver,
-  eventSource,
-  eventFns = getEventFnsBasedOnEventSource({ eventSource })
-}) {
-  setCurrentLambdaInvoke({ event, context })
-  try {
-    const requestOptions = eventFns.request({
-      event,
-      socketPath: getSocketPath({ socketPathSuffix: server._socketPathSuffix })
-    })
-    const req = http.request(requestOptions, (response) => forwardResponse({ server, response, resolver, responseFn: eventFns.response }))
-
-    if (event.body) {
-      const body = getEventBody({ event })
-      req.write(body)
-    }
-
-    req.on('error', (error) => forwardConnectionErrorResponseToApiGateway({ error, resolver, responseFn: eventFns.response }))
-      .end()
-  } catch (error) {
-    forwardLibraryErrorResponseToApiGateway({ error, resolver })
-    return server
-  }
-}
-
-function startServer ({ server }) {
-  return server.listen(getSocketPath({ socketPathSuffix: server._socketPathSuffix }))
-}
-
-function getSocketPath ({ socketPathSuffix }) {
-  /* only running tests on Linux; Window support is for local dev only */
-  /* istanbul ignore if */
-  if (/^win/.test(process.platform)) {
-    const path = require('path')
-    return path.join('\\\\?\\pipe', process.cwd(), `server-${socketPathSuffix}`)
-  } else {
-    return `/tmp/server-${socketPathSuffix}.sock`
-  }
-}
-
-function getRandomString () {
-  return Math.random().toString(36).substring(2, 15)
 }
 
 function createServer ({
@@ -256,16 +64,6 @@ function createServer ({
   return server
 }
 
-function getEventSourceBasedOnEvent ({
-  event
-}) {
-  if (event && event.requestContext && event.requestContext.elb) return 'ALB'
-  if (event && event.requestContext && event.requestContext.stage) return 'API_GATEWAY'
-  if (event && event.Records) return 'LAMBDA_EDGE'
-
-  throw new Error('Unable to determine event source based on event.')
-}
-
 function proxy ({
   server,
   event = {},
@@ -275,6 +73,7 @@ function proxy ({
   eventSource = getEventSourceBasedOnEvent({ event }),
   eventFns = getEventFnsBasedOnEventSource({ eventSource })
 }) {
+  setCurrentLambdaInvoke({ event, context })
   return {
     server,
     promise: new Promise((resolve, reject) => {
@@ -310,23 +109,6 @@ function proxy ({
           }))
       }
     })
-  }
-}
-
-function makeResolver ({
-  context,
-  callback,
-  promise,
-  resolutionMode
-}) {
-  return {
-    succeed: ({
-      response
-    }) => {
-      if (resolutionMode === 'CONTEXT_SUCCEED') return context.succeed(response)
-      if (resolutionMode === 'CALLBACK') return callback(null, response)
-      if (resolutionMode === 'PROMISE') return promise.resolve(response)
-    }
   }
 }
 
@@ -380,18 +162,7 @@ function configure ({
   }
 }
 
-exports.configure = configure
-exports.getCurrentLambdaInvoke = getCurrentLambdaInvoke
-
-/* istanbul ignore else */
-if (process.env.NODE_ENV === 'test') {
-  exports.getPathWithQueryStringParams = getPathWithQueryStringParams
-  exports.mapApiGatewayEventToHttpRequest = mapApiGatewayEventToHttpRequest
-  exports.forwardResponseToApiGateway = forwardResponseToApiGateway
-  exports.forwardConnectionErrorResponseToApiGateway = forwardConnectionErrorResponseToApiGateway
-  exports.forwardLibraryErrorResponseToApiGateway = forwardLibraryErrorResponseToApiGateway
-  exports.forwardRequestToNodeServer = forwardRequestToNodeServer
-  exports.startServer = startServer
-  exports.getSocketPath = getSocketPath
-  exports.makeResolver = makeResolver
+module.exports = {
+  configure,
+  getCurrentLambdaInvoke
 }
